move tests to CandidateTest
